Poner los avisos

+++ src/main/java/epsum/curso/conexiondatos/ventanas/PanelEmpleado.java
@@ -18,6 +18,8 @@ import javax.swing.border.Border;
 import javax.swing.table.DefaultTableModel;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.DuplicateKeyException;
 
 import epsum.curso.conexiondatos.servicios.DatoLaboralService;
 import epsum.curso.conexiondatos.servicios.DatosPersonalesService;
@@ -30,6 +32,8 @@ import epsum.curso.conexiondatos.servicios.EmpresaService;
 import lombok.Data;
 @Data
 public class PanelEmpleado extends PanelComponente {
+	private static final String String = null;
+
 	public PanelEmpleado(Object[] cabeceras, Object[][] datos, String titulo) {
 		super(cabeceras, datos, titulo);
 	}
@@ -56,6 +60,8 @@ public class PanelEmpleado extends PanelComponente {
 	
 	@Override
 	public void alta() {
+		 int confirmation = JOptionPane.showConfirmDialog(null, "¿Deseas agregar un nuevo registro?", "Confirmación", JOptionPane.YES_NO_OPTION);
+		    if (confirmation == JOptionPane.YES_OPTION) {
 		
 		DefaultTableModel defaultTableModel = (DefaultTableModel) getTabla().getModel();
   
@@ -91,37 +97,63 @@ public class PanelEmpleado extends PanelComponente {
         defaultTableModel.addRow(datos);
 
       getTabla().getColumnModel().getColumn(7).setCellEditor(new DefaultCellEditor(jComboBoxDatosLaborales));
-      
+		    }
 	}
 	
 
 	@Override
 	public void baja() {
 		int id = Integer.parseInt((String) getTabla().getModel().getValueAt(getTabla().getSelectedRow(), 0));
-		getEmpleadoService().deleteById(id);
-		 DefaultTableModel defaultTableModel=(DefaultTableModel) getTabla().getModel();
-		defaultTableModel.removeRow(getTabla().getSelectedRow());
-		
+	    int confirmation = JOptionPane.showConfirmDialog(null, "¿Estás seguro?", "Confirmación", JOptionPane.YES_NO_OPTION);
+	    if (confirmation == JOptionPane.YES_OPTION) {
+
+	        try {
+				getEmpleadoService().deleteById(id);
+				JOptionPane.showMessageDialog(null, "Registro borrado correctamente", "Éxito", JOptionPane.INFORMATION_MESSAGE);
+				 DefaultTableModel defaultTableModel = (DefaultTableModel) getTabla().getModel();
+			     defaultTableModel.removeRow(getTabla().getSelectedRow());
+			} catch (Exception e) {	
+				e.printStackTrace();
+				JOptionPane.showMessageDialog(null, "Registro no se ha podido borrar", "Éxito", JOptionPane.INFORMATION_MESSAGE);
+			} 
+
+	       
+	    }
 	}
 
 	@Override
 	public void modificar() {
-		int fila = 0;
-		for (int i = 0; i <getTabla().getModel().getRowCount(); i++) {
-			Empleado empleado = new Empleado();
-			empleado.setId((int)getTabla().getModel().getValueAt(i,0));
-			empleado.setNombre((String) getTabla().getModel().getValueAt(i, 1));
-			empleado.setDni((String) getTabla().getModel().getValueAt(i, 2));
-			empleado.setEmail((String) getTabla().getModel().getValueAt(i, 3));
-			empleado.setTelefono((String) getTabla().getModel().getValueAt(i, 4));
-			empleado.setEmpresa((Empresa) getTabla().getModel().getValueAt(i, 5));
-			empleado.setDatoLaboral((DatoLaboral) getTabla().getModel().getValueAt(i, 6));
-			empleado.setDatoPersonal((DatoPersonal) getTabla().getModel().getValueAt(i, 7));
-			getEmpleadoService().save(empleado);
-			getTabla().getModel().setValueAt(String.valueOf(empleado.getId()),fila, 0);
-			fila++;
-			
-		}
+		try {
+	        for (int i = 0; i < getTabla().getModel().getRowCount(); i++) {
+	            Empleado empleado = new Empleado();
+	            empleado.setId(Integer.parseInt((String) getTabla().getModel().getValueAt(i, 0)));
+	            empleado.setNombre((String) getTabla().getModel().getValueAt(i, 1));
+	            empleado.setDni((String) getTabla().getModel().getValueAt(i, 2));
+	            empleado.setEmail((String) getTabla().getModel().getValueAt(i, 3));
+	            empleado.setTelefono((String) getTabla().getModel().getValueAt(i, 4));
+	            empleado.setEmpresa((Empresa) getTabla().getModel().getValueAt(i, 5));
+	            empleado.setDatoLaboral((DatoLaboral) getTabla().getModel().getValueAt(i, 6));
+	            empleado.setDatoPersonal((DatoPersonal) getTabla().getModel().getValueAt(i, 7));
+	            
+	         //Aqui intenta guardarlo:
+	            getEmpleadoService().save(empleado);
+	            DefaultTableModel defaultTableModel= (DefaultTableModel) getTabla().getModel();
+	            defaultTableModel.setValueAt(String.valueOf(empleado.getId()), i, 0);
+	        }
+	        JOptionPane.showMessageDialog(null, "Las modificaciones fueron exitosas", "Éxito", JOptionPane.INFORMATION_MESSAGE);
+	    
+	    } catch (DuplicateKeyException e) {
+	        JOptionPane.showMessageDialog(null, "Error: Clave duplicada. No se puede guardar el cargo.", "Error", JOptionPane.ERROR_MESSAGE);
+	        e.printStackTrace(); 
+	        
+	    } catch (DataIntegrityViolationException e) {
+	        JOptionPane.showMessageDialog(null, "Error: Violación de integridad. Verifique los datos.", "Error", JOptionPane.ERROR_MESSAGE);
+	        e.printStackTrace(); 
+	        
+	    }catch (Exception e) {
+	        JOptionPane.showMessageDialog(null, "Hubo un problema al modificar los datos", "Error", JOptionPane.ERROR_MESSAGE);
+	        e.printStackTrace(); 
+	    }
 	}
 
 }
